requirements.txt
aiogram_dialog
asyncpg
python-dotenv
redis
tortoise-api-model
